requirements.txt
jaxtyping
tqdm
jax[cpu]
pytest
plum-dispatch
simple-pytree
seaborn
hdf5storage
ipykernel
jupyter
numpy
pandas